Separate test file into 4 different files.
